MyBatis Updates.

Approved-by: Thiru Pathi

+++ src/Service/ProductServiceImplementation.java
@@ -3,31 +3,30 @@ package Service;
 import DAO.*;
 import Entity.Product;
 
-import java.sql.SQLException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 
 public class ProductServiceImplementation implements ProductService {
 
-	private final ProductDAO productDAO = new ProductDAOImplementation();
-	private final String NAME_REGEX = "^[a-zA-Z\\s]{1,30}$";
+	private final ProductDAO productDAO = new ProductDAO();
+	private final String NAME_REGEX = "^[a-zA-Z\\s]{3,30}$";
 	private final String CODE_REGEX = "^[a-zA-Z0-9]{2,6}$";
 	private final String UNIT_CODE_REGEX = "^[a-zA-Z]{1,4}$";
 
-	public Product create(Product product) throws SQLException, ApplicationErrorException, UniqueConstraintException, UnitCodeViolationException, InvalidTemplateException {
+	public Product create(Product product) throws Exception {
 		validate(product);
 		return productDAO.create(product);
 	}
 
-	public Integer count() throws ApplicationErrorException {
-		return productDAO.count();
+	public Integer count(String attribute, String searchText) throws ApplicationErrorException {
+			return productDAO.count(attribute, searchText);
 	}
 
 	public List<Product> list(HashMap<String, String> listattributes) throws ApplicationErrorException, PageCountOutOfBoundsException {
 		List<Product> productList;
 		if(Collections.frequency(listattributes.values(), null) == listattributes.size() - 1 && listattributes.get("Searchtext") != null) {
-			productList = productDAO.list(listattributes.get("Searchtext"));
+			productList = productDAO.searchList(listattributes.get("Searchtext"));
 		} else {
 			int pageLength = Integer.parseInt(listattributes.get("Pagelength"));
 			int pageNumber = Integer.parseInt(listattributes.get("Pagenumber"));
@@ -37,13 +36,16 @@ public class ProductServiceImplementation implements ProductService {
 		return productList;
 	}
 
-	public Product edit(Product product) throws SQLException, ApplicationErrorException, UniqueConstraintException, UnitCodeViolationException, InvalidTemplateException {
+	public Product edit(Product product) throws Exception {
 		validate(product);
 		return productDAO.edit(product);
 	}
 
 	public Integer delete(String parameter) throws ApplicationErrorException {
-		return productDAO.delete(parameter);
+		if(parameter != null)
+			return productDAO.delete(parameter);
+		else
+			return -1;
 	}
 
 	/**
@@ -52,6 +54,8 @@ public class ProductServiceImplementation implements ProductService {
 	 * @param product Product to be Validated
 	 */
 	private void validate(Product product) throws InvalidTemplateException {
+		if(product == null)
+			throw new NullPointerException(">> Product cannot be Null!!");
 		if(product.getCode() != null && ! product.getCode().matches(CODE_REGEX))
 			throw new InvalidTemplateException(">> Invalid Product Code!!");
 		if(product.getName() != null && ! product.getName().matches(NAME_REGEX))
