+++ src/Service/PurchaseServiceImplementation.java
@@ -5,29 +5,29 @@ import Entity.Product;
 import Entity.Purchase;
 import Entity.PurchaseItem;
 
-import java.sql.SQLException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 
 public class PurchaseServiceImplementation implements PurchaseService {
 
-	private final PurchaseDAO purchaseDAO = new PurchaseDAOImplementation();
+	private final PurchaseDAO purchaseDAO = new PurchaseDAO();
+	private final String dateRegex = "([12]\\d{3}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01]))";
 
 
 	@Override
-	public Purchase create(Purchase purchase) throws ApplicationErrorException, SQLException, UnDividableEntityException {
-		ProductDAO productDAO = new ProductDAOImplementation();
-		UnitDAO unitDAO = new UnitDAOImplementation();
+	public Purchase create(Purchase purchase) throws Exception {
+		ProductDAO productDAO = new ProductDAO();
+		UnitDAO unitDAO = new UnitDAO();
 		boolean isDividable;
+		validate(purchase);
 		for(PurchaseItem purchaseItem: purchase.getPurchaseItemList()) {
 			try {
 				Product product = productDAO.findByCode(purchaseItem.getProduct().getCode());
-				if(product != null)
-					purchaseItem.setProduct(product);
+				if(product != null) purchaseItem.setProduct(product);
 				isDividable = unitDAO.findByCode(product.getunitcode()).getIsDividable();
 			} catch(NullPointerException e) {
-				throw new ApplicationErrorException(">> Product code '"+purchaseItem.getProduct().getCode() + "' does not exists!");
+				throw new ApplicationErrorException(">> Product code '" + purchaseItem.getProduct().getCode() + "' does not exists!");
 			}
 			if(! isDividable && purchaseItem.getQuantity() % 1 != 0) {
 				throw new UnDividableEntityException(">> Product code '" + purchaseItem.getProduct().getCode() + "' is not of dividable unit!");
@@ -38,24 +38,27 @@ public class PurchaseServiceImplementation implements PurchaseService {
 
 
 	@Override
-	public Integer count(String parameter) throws ApplicationErrorException {
-		String dateRegex = "([12]\\d{3}-(0[1-9]|1[0-2])-(0[1-9]|[12]\\d|3[01]))";
-		if(parameter != null) {
-			if(! parameter.matches(dateRegex)) return - 1;
+	public Integer count(String attribute, String searchText) throws ApplicationErrorException, InvalidTemplateException {
+		if(searchText != null) {
+			if(!searchText.matches(dateRegex))
+				throw new InvalidTemplateException(">> Invalid Date format!! Must be in YYYY-MM-DD format!");
 		}
-		return purchaseDAO.count(parameter);
+		return purchaseDAO.count(attribute, searchText);
 	}
 
 
 	@Override
-	public List<Purchase> list(HashMap<String, String> listattributes) throws ApplicationErrorException {
+	public List<Purchase> list(HashMap<String, String> listattributes) throws ApplicationErrorException, InvalidTemplateException {
 		List<Purchase> purchaseList;
 		if(Collections.frequency(listattributes.values(), null) == listattributes.size() - 1 && listattributes.get("Searchtext") != null) {
-			purchaseList = purchaseDAO.list(listattributes.get("Searchtext"));
+			purchaseList = purchaseDAO.searchList(listattributes.get("Searchtext"));
 		} else {
 			int pageLength = Integer.parseInt(listattributes.get("Pagelength"));
 			int pageNumber = Integer.parseInt(listattributes.get("Pagenumber"));
 			int offset = (pageLength * pageNumber) - pageLength;
+			if(listattributes.get("Attribute").equals("date"))
+				if(! (listattributes.get("Searchtext").replace("'", "").matches(dateRegex)))
+					throw new InvalidTemplateException(">> Invalid Format for Attribute date!! Must be in format YYYY-MM-DD");
 			purchaseList = purchaseDAO.list(listattributes.get("Attribute"), listattributes.get("Searchtext"), pageLength, offset);
 		}
 		return purchaseList;
@@ -64,6 +67,16 @@ public class PurchaseServiceImplementation implements PurchaseService {
 
 	@Override
 	public Integer delete(String invoice) throws ApplicationErrorException {
-		return purchaseDAO.delete(Integer.parseInt(invoice));
+		if(invoice != null)
+			return purchaseDAO.delete(Integer.parseInt(invoice));
+		else
+			return -1;
+	}
+
+	private void validate(Purchase purchase) throws InvalidTemplateException {
+		if(purchase == null)
+			throw new NullPointerException(">> Purchase cannot be Null");
+		if(purchase.getDate() != null && ! purchase.getDate().matches(dateRegex))
+			throw new InvalidTemplateException(">> Date Format is Invalid!! Must be YYYY-MM-DD!!");
 	}
 }
