+++ src/Service/UnitServiceImplementation.java
@@ -1,22 +1,18 @@
 package Service;
 
-import DAO.ApplicationErrorException;
-import DAO.UniqueConstraintException;
-import DAO.UnitDAO;
-import DAO.UnitDAOImplementation;
+import DAO.*;
 import Entity.Unit;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class UnitServiceImplementation implements UnitService {
-	private final UnitDAO unitDAO = new UnitDAOImplementation();
-	private final String NAME_REGEX = "^[a-zA-Z\\s]{1,30}$";
+	private final UnitDAO unitDAO = new UnitDAO();
+	private final String NAME_REGEX = "^[a-zA-Z\\s]{3,30}$";
 	private final String CODE_REGEX = "^[a-zA-Z]{1,4}$";
 
 
 	@Override
-	public Unit create(Unit unit) throws SQLException, ApplicationErrorException, UniqueConstraintException, InvalidTemplateException {
+	public Unit create(Unit unit) throws Exception {
 		validate(unit);
 		return unitDAO.create(unit);
 	}
@@ -29,15 +25,18 @@ public class UnitServiceImplementation implements UnitService {
 
 
 	@Override
-	public Unit edit(Unit unit) throws SQLException, ApplicationErrorException, UniqueConstraintException, InvalidTemplateException {
+	public Unit edit(Unit unit) throws Exception {
 		validate(unit);
 		return unitDAO.edit(unit);
 	}
 
 
 	@Override
-	public Integer delete(String code) throws ApplicationErrorException {
-		return unitDAO.delete(code);
+	public Integer delete(String code) throws Exception {
+		if(code != null)
+			return unitDAO.delete(code);
+		else
+			return -1;
 	}
 
 
@@ -47,6 +46,8 @@ public class UnitServiceImplementation implements UnitService {
 	 * @param unit Unit to be validated
 	 */
 	private void validate(Unit unit) throws InvalidTemplateException {
+		if(unit == null)
+			throw new NullPointerException(">> Unit Cannot be Null!!");
 		if(unit.getName() != null && ! unit.getName().matches(NAME_REGEX))
 			throw new InvalidTemplateException(">> Invalid Unit Name!!");
 		if(unit.getCode() != null && ! unit.getCode().matches(CODE_REGEX))
